modified:   validated/libft/ft_atoi.c 	modified:   validated/libft/ft_itoa.c 	modified:   validated/libft/ft_split.c 	modified:   validated/libft/ft_strtrim.c 	modified:   validated/libft/ft_substr.c 	modified:   validated/libft/libft.h

+++ validated/libft/ft_atoi.c
@@ -6,11 +6,11 @@
 /*   By: achauvie <achauvie@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/10/13 15:26:33 by achauvie          #+#    #+#             */
-/*   Updated: 2025/10/22 09:39:37 by achauvie         ###   ########.fr       */
+/*   Updated: 2025/11/14 08:57:40 by achauvie         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-int	ft_atoi(const char *nptr)
+long	ft_atoi(const char *nptr)
 {
 	int	i;
 	int	nb;
